♻ Repurpose interceptor as page network layer
Iterception disabled by default & allows a nicer `page.network.idle()` syntax
